Extract ActButton
